Add tool to produce pickled data
The generators produce their results in SQLite database,
which is convenient for browsing, fetching and updating
as persistent storage, however for the real-time usage
in Streamlit application the pickled DataFrame object
will be more convenient. This commit introduces a tool
to produce such, that also adds some additional columns
computed on the source data.