Hypenation rework
Removing the hypenation causes issues in some very specific cases,
such as in bibliography and in places with a lot of long words,
making LaTeX placing the words on margin. On the other hand,
with default hypenation rules, LaTeX tends to split too many words,
where the justified text with a little more spacing look better,
at least to my taste. The solution is to allow hypenation,
but increase the default penalty, so LaTeX would split the word
only if there is no really other option to place it nicely.